+++ src/test/java/wooteco/subway/path/domain/fare/ageStrategy/AgeAppliedRuleTest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.path.domain;
+package wooteco.subway.path.domain.fare.ageStrategy;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import wooteco.subway.path.domain.fare.ageStrategy.AgeAppliedRule;
 
 public class AgeAppliedRuleTest {
 
@@ -22,7 +23,7 @@ public class AgeAppliedRuleTest {
         return Stream.of(
             Arguments.of(3, AgeAppliedRule.INFANT),
             Arguments.of(9, AgeAppliedRule.CHILD),
-            Arguments.of(15, AgeAppliedRule.JUVENILLE),
+            Arguments.of(15, AgeAppliedRule.JUVENILE),
             Arguments.of(20, AgeAppliedRule.OTHER)
         );
     }
